[SADDY-QT-WIDGET] Moved widget in library.
