+++ include/qt/sadqtrenderer.h
