+++ plugins/qt-widget/sadqtmainloop.cpp
@@ -1,4 +1,4 @@
-#include "sadqtmainloop.h"
+#include "qt/sadqtmainloop.h"
 
 #include <renderer.h>
 #include <window.h>
