+++ plugins/qt-widget/sadqtwindow.cpp
@@ -1,6 +1,6 @@
-#include "sadqtwindow.h"
-#include "sadqtrenderer.h"
-#include "sadqtopenglwidget.h"
+#include "qt/sadqtwindow.h"
+#include "qt/sadqtrenderer.h"
+#include "qt/sadqtopenglwidget.h"
 
 #include <QTimer>
